requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.11.0; python_version >= '3.9'
asgiref==3.9.2; python_version >= '3.9'
certifi==2025.8.3; python_version >= '3.7'
distro==1.9.0; python_version >= '3.6'
dj-database-url==3.0.1
django==5.2.7; python_version >= '3.10'
django-jet-reboot==1.3.10
h11==0.16.0; python_version >= '3.8'
httpcore==1.0.9; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
jiter==0.11.0; python_version >= '3.9'
openai==2.1.0; python_version >= '3.8'
pillow==11.3.0; python_version >= '3.9'
psycopg2-binary==2.9.10; python_version >= '3.8'
pydantic==2.11.9; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
python-dotenv==1.1.1; python_version >= '3.9'
sniffio==1.3.1; python_version >= '3.7'
sqlparse==0.5.3; python_version >= '3.8'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.15.0; python_version >= '3.9'
typing-inspection==0.4.2; python_version >= '3.9'
whitenoise==6.11.0; python_version >= '3.9'
